requirements.txt
azure-core==1.30.1
azure-identity==1.16.0
azure-search-documents==11.4.0
semantic-kernel==0.9.5b1
openai==1.17.1
langchain==0.1.16
langchain-openai==0.1.3
tiktoken==0.6.0
python-dotenv==1.0.1
requests==2.31.0
unstructured==0.13.2
markdown==3.6
qdrant-client==1.8.2
chromadb==0.4.24
yfinance==0.2.37
langchain-community==0.0.32
pymongo==4.6.3
